requirements.txt
python-dotenv==0.20.0
pyyaml==6.0